@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "camel-ai"
-version = "0.1.6.0"
+version = "0.1.6.9"
 authors = ["CAMEL-AI.org"]
 description = "Communicative Agents for AI Society Study"
 readme = "README.md"
@@ -27,7 +27,7 @@ repository = "https://github.com/camel-ai/camel"
 documentation = "https://docs.camel-ai.org"
 
 [tool.poetry.dependencies]
-python = ">=3.9.0,<3.12"
+python = ">=3.10.0,<3.12"
 numpy = "^1"
 openai = "^1.2.3"
 groq = "^0.5.0"
@@ -49,8 +49,8 @@ pillow = "^10.2.0"
 # model-platforms
 litellm = { version = "^1.38.1", optional = true }
 google-generativeai = { version = "^0.6.0", optional = true }
-mistralai = { version = "^0.4.2", optional = true }
-mistral-common = { version = "^1.3.3", optional = true }
+mistralai = { version = "^1.0.0", optional = true }
+reka-api = { version = "^3.0.8", optional = true } 
 
 # huggingface-agent
 transformers = { version = "^4", optional = true }
@@ -75,7 +75,7 @@ googlemaps = { version = "^4.10.0", optional = true }
 requests_oauthlib = { version = "^1.3.1", optional = true }
 prance = { version = "^23.6.21.0", optional = true }
 openapi-spec-validator = { version = "^0.7.1", optional = true }
-unstructured = { extras = ["all-docs"], version = "^0.10.30", optional = true }
+unstructured = { extras = ["all-docs"], version = "^0.10", optional = true }
 firecrawl-py = { version = "^0.0.20", optional = true }
 slack-sdk = { version = "^3.27.2", optional = true }
 pydub = { version = "^0.25.1", optional = true }
@@ -84,6 +84,11 @@ imageio = { extras = ["pyav"], version = "^2.34.2" }
 pyTelegramBotAPI = { version = "^4.18.0", optional = true }
 "discord.py" = { version = "^2.3.2", optional = true }
 docker = { version = "^7.1.0", optional = true }
+agentops = { version = "^0.3.6", optional = true }
+azure-storage-blob = { version = "^12.21.0", optional = true }
+google-cloud-storage = { version = "^2.18.0", optional = true }
+botocore = { version = "^1.35.3", optional = true }
+nltk = { version = "3.8.1", optional = true }
 
 # encoders
 sentence-transformers = { version = "^3.0.1", optional = true }
@@ -114,7 +119,7 @@ model-platforms = [
     "litellm",
     "google-generativeai",
     "mistralai",
-    "mistral-common",
+    "reka-api",
 ]
 
 huggingface-agent = [
@@ -147,6 +152,7 @@ tools = [
     "openapi-spec-validator",
     "unstructured",
     "firecrawl-py",
+    "nltk",
     "imageio",
     "pillow",
     "slack-sdk",
@@ -156,6 +162,7 @@ tools = [
     "discord.py",
     "docker",
     "jupyter_client",
+    "agentops",
 ]
 
 vector-databases = [
@@ -169,6 +176,12 @@ graph-storages = [
 
 kv-stroages = [
     "redis",
+]
+
+object-storages = [
+    "azure-storage-blob",
+    "google-cloud-storage",
+    "botocore",
 ]
 
 retrievers = [
@@ -203,6 +216,7 @@ all = [
     "prance",
     "openapi-spec-validator",
     "unstructured",
+    "nltk",
     "firecrawl-py",
     "imageio",
     "slack-sdk",
@@ -210,6 +224,7 @@ all = [
     "pillow",
     "docker",
     "jupyter_client",
+    "agentops",
     # vector-database
     "qdrant-client",
     "pymilvus",
@@ -224,10 +239,14 @@ all = [
     # model platforms
     "litellm",
     "mistralai",
-    "mistral-common",
     "google-generativeai",
+    "reka-api",
     # kv-storages
     "redis",
+    # object-storages
+    "azure-storage-blob",
+    "google-cloud-storage",
+    "botocore",
 ]
 
 [tool.poetry.group.dev]
@@ -342,6 +361,7 @@ module = [
     "fitz",
     "qdrant_client.*",
     "unstructured.*",
+    "nltk",
     "firecrawl",
     "imageio",
     "rank_bm25",
@@ -359,6 +379,8 @@ module = [
     "google.*",
     "google_generativeai",
     "mistralai",
-    "mistral_common",
+    "reka-api",
+    "agentops",
+    "botocore.*"
 ]
 ignore_missing_imports = true
