@@ -12,14 +12,13 @@
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 
-from .base_io import File, read_file
+from .base_io import File
 from .firecrawl_reader import Firecrawl
 from .jina_url_reader import JinaURLReader
 from .unstructured_io import UnstructuredIO
 
 __all__ = [
     'File',
-    'read_file',
     'UnstructuredIO',
     'JinaURLReader',
     'Firecrawl',
