@@ -51,21 +51,21 @@ class GroqModel(BaseModelBackend):
             api_key (Optional[str]): The API key for authenticating with the
                 Groq service. (default: :obj:`None`).
             url (Optional[str]): The url to the Groq service. (default:
-                :obj:`None`)
+                :obj:`"https://api.groq.com/openai/v1"`)
             token_counter (Optional[BaseTokenCounter]): Token counter to use
                 for the model. If not provided, `OpenAITokenCounter(ModelType.
-                GPT_3_5_TURBO)` will be used.
+                GPT_4O_MINI)` will be used.
         """
         super().__init__(
             model_type, model_config_dict, api_key, url, token_counter
         )
-        self._url = url or "https://api.groq.com/openai/v1"
+        self._url = url or os.environ.get("GROQ_API_BASE_URL")
         self._api_key = api_key or os.environ.get("GROQ_API_KEY")
         self._client = OpenAI(
             timeout=60,
             max_retries=3,
             api_key=self._api_key,
-            base_url=self._url,
+            base_url=self._url or "https://api.groq.com/openai/v1",
         )
         self._token_counter = token_counter
 
@@ -80,7 +80,7 @@ class GroqModel(BaseModelBackend):
         # Make sure you have the access to these open-source model in
         # HuggingFace
         if not self._token_counter:
-            self._token_counter = OpenAITokenCounter(ModelType.GPT_3_5_TURBO)
+            self._token_counter = OpenAITokenCounter(ModelType.GPT_4O_MINI)
         return self._token_counter
 
     @api_keys_required("GROQ_API_KEY")
