@@ -22,8 +22,12 @@ from camel.models.litellm_model import LiteLLMModel
 from camel.models.mistral_model import MistralModel
 from camel.models.ollama_model import OllamaModel
 from camel.models.open_source_model import OpenSourceModel
+from camel.models.openai_compatibility_model import OpenAICompatibilityModel
 from camel.models.openai_model import OpenAIModel
+from camel.models.reka_model import RekaModel
+from camel.models.samba_model import SambaModel
 from camel.models.stub_model import StubModel
+from camel.models.togetherai_model import TogetherAIModel
 from camel.models.vllm_model import VLLMModel
 from camel.models.zhipuai_model import ZhipuAIModel
 from camel.types import ModelPlatformType, ModelType
@@ -90,6 +94,10 @@ class ModelFactory:
                 model_class = GeminiModel
             elif model_platform.is_mistral and model_type.is_mistral:
                 model_class = MistralModel
+            elif model_platform.is_reka and model_type.is_reka:
+                model_class = RekaModel
+            elif model_platform.is_samba and model_type.is_samba:
+                model_class = SambaModel
             elif model_type == ModelType.STUB:
                 model_class = StubModel
             else:
@@ -105,11 +113,15 @@ class ModelFactory:
                 )
             elif model_platform.is_vllm:
                 model_class = VLLMModel
-                return model_class(
-                    model_type, model_config_dict, url, api_key, token_counter
-                )
             elif model_platform.is_litellm:
                 model_class = LiteLLMModel
+            elif model_platform.is_openai_compatibility_model:
+                model_class = OpenAICompatibilityModel
+            elif model_platform.is_together:
+                model_class = TogetherAIModel
+                return model_class(
+                    model_type, model_config_dict, api_key, token_counter
+                )
             else:
                 raise ValueError(
                     f"Unknown pair of model platform `{model_platform}` "
