@@ -486,7 +486,15 @@ class RolePlaying:
                 ),
             )
         user_msg = self._reduce_message_options(user_response.msgs)
-        self.user_agent.record_message(user_msg)
+
+        # To prevent recording the same memory more than once (once in chat
+        # step and once in role play), and the model generates only one
+        # response when multi-response support is enabled.
+        if (
+            'n' in self.user_agent.model_config_dict.keys()
+            and self.user_agent.model_config_dict['n'] > 1
+        ):
+            self.user_agent.record_message(user_msg)
 
         assistant_response = self.assistant_agent.step(user_msg)
         if assistant_response.terminated or assistant_response.msgs is None:
@@ -501,7 +509,15 @@ class RolePlaying:
                 ),
             )
         assistant_msg = self._reduce_message_options(assistant_response.msgs)
-        self.assistant_agent.record_message(assistant_msg)
+
+        # To prevent recording the same memory more than once (once in chat
+        # step and once in role play), and the model generates only one
+        # response when multi-response support is enabled.
+        if (
+            'n' in self.assistant_agent.model_config_dict.keys()
+            and self.assistant_agent.model_config_dict['n'] > 1
+        ):
+            self.assistant_agent.record_message(assistant_msg)
 
         return (
             ChatAgentResponse(
