@@ -18,11 +18,10 @@ from typing import Any, Callable, Dict, List, Optional, Tuple
 import requests
 
 from camel.toolkits import OpenAIFunction, openapi_security_config
-from camel.toolkits.base import BaseToolkit
 from camel.types import OpenAPIName
 
 
-class OpenAPIToolkit(BaseToolkit):
+class OpenAPIToolkit:
     r"""A class representing a toolkit for interacting with OpenAPI APIs.
 
     This class provides methods for interacting with APIs based on OpenAPI
