@@ -26,7 +26,9 @@ from PIL import Image
 from camel.types import ModelType, OpenAIImageType, OpenAIVisionDetailType
 
 if TYPE_CHECKING:
-    from mistral_common.protocol.instruct.request import ChatCompletionRequest
+    from mistral_common.protocol.instruct.request import (  # type:ignore[import-not-found]
+        ChatCompletionRequest,
+    )
 
     from camel.messages import OpenAIMessage
 
@@ -517,7 +519,9 @@ class MistralTokenCounter(BaseTokenCounter):
             model_type (ModelType): Model type for which tokens will be
                 counted.
         """
-        from mistral_common.tokens.tokenizers.mistral import MistralTokenizer
+        from mistral_common.tokens.tokenizers.mistral import (  # type:ignore[import-not-found]
+            MistralTokenizer,
+        )
 
         self.model_type = model_type
 
@@ -565,7 +569,7 @@ class MistralTokenCounter(BaseTokenCounter):
         """
 
         from mistral_common.protocol.instruct.request import (
-            ChatCompletionRequest,
+            ChatCompletionRequest,  # type:ignore[import-not-found]
         )
 
         mistral_request = ChatCompletionRequest(  # type: ignore[type-var]
