@@ -20,7 +20,9 @@ from camel.utils import dependencies_required
 
 # Conditionally import telebot types only for type checking
 if TYPE_CHECKING:
-    from telebot.types import Message  # type: ignore[import-untyped]
+    from telebot.types import (
+        Message,  # type: ignore[import-not-found]
+    )
 
 
 class TelegramBot:
@@ -49,7 +51,7 @@ class TelegramBot:
         else:
             self.token = telegram_token
 
-        import telebot  # type: ignore[import-untyped]
+        import telebot  # type: ignore[import-not-found]
 
         self.bot = telebot.TeleBot(token=self.token)
 
