@@ -119,7 +119,7 @@ def solve_issue(
     ).as_dict()
 
     model = ModelFactory.create(
-        model_platform=ModelPlatformType.OpenAI,
+        model_platform=ModelPlatformType.OPENAI,
         model_type=ModelType.GPT_3_5_TURBO,
         model_config_dict=assistant_model_config_dict,
     )
