@@ -119,7 +119,7 @@ def translate_content(
             model_config = ChatGPTConfig(stream=True)
 
         model = ModelFactory.create(
-            model_platform=ModelPlatformType.OpenAI,
+            model_platform=ModelPlatformType.OPENAI,
             model_type=ModelType.GPT_3_5_TURBO,
             model_config=model_config,
         )
